FIX: DO NOT clear logInstance cache..
This is a static class - singleton - if you create a logger with the 'old' interface you get a raw-pointer but we still manage the instances through shared_ptr. This is causing a problem IF we clear the logInstance cache. Since the raw-ptr version won't be tracked these loggers will go out...